Remove src/preload and move sentry init
