+++ src/v2/render.tsx
@@ -5,7 +5,7 @@ import * as DOM from "react-dom";
 import "core-js";
 import "core-js/proposals/array-find-from-last";
 import "remove-focus-outline";
-import { initializeSentry } from "../preload/sentry";
+import { initializeSentry } from "./utils/sentry";
 import App from "./App";
 
 import { getCurrentWindow } from "@electron/remote";
